- Recyclerview adapter 기본 구성 - ViewModel 구성 - Data 오류 수정 - 패키지 구성

+++ app/src/main/java/com/coooldoggy/itunesmusic/framework/data/Song.kt
@@ -49,10 +49,10 @@ data class Song(
     val artworkUrl100: String,
 
     @SerializedName("collectionPrice")
-    val collectionPrice: Int,
+    val collectionPrice: Float,
 
     @SerializedName("trackPrice")
-    val trackPrice: Int,
+    val trackPrice: Float,
 
     @SerializedName("releaseDate")
     val releaseDate: String,
